Moved Overlay rendering parts to a separate class in PrioritySigns tool
