+++ TLM/TLM/UI/SubTools/PrioritySigns/PrioritySignsTool.cs
@@ -1,9 +1,8 @@
-namespace TrafficManager.UI.SubTools {
+namespace TrafficManager.UI.SubTools.PrioritySigns {
+    using System;
+    using System.Collections.Generic;
     using ColossalFramework;
     using CSUtil.Commons;
-    using static Util.SegmentTraverser;
-    using System.Collections.Generic;
-    using System;
     using TrafficManager.API.Manager;
     using TrafficManager.API.Traffic.Data;
     using TrafficManager.API.Traffic.Enums;
@@ -12,6 +11,7 @@ namespace TrafficManager.UI.SubTools {
     using TrafficManager.UI.Textures;
     using TrafficManager.Util;
     using UnityEngine;
+    using static TrafficManager.Util.SegmentTraverser;
 
     public class PrioritySignsTool : LegacySubTool {
         private enum PrioritySignsMassEditMode {
@@ -160,66 +160,6 @@ namespace TrafficManager.UI.SubTools {
         public override void OnToolGUI(Event e) { }
 
         /// <summary>
-        /// Thread safe handling of mass edit overlay.
-        /// </summary>
-        public static class MassEditOVerlay {
-            private static object _lock = new object();
-
-            private static bool _show = false;
-            public static bool Show {
-                get {
-                    lock (_lock) {
-                        return _show;
-                    }
-                }
-                set {
-                    lock (_lock) {
-                        _show = value;
-                    }
-                }
-            }
-
-            private static DateTime _timer = DateTime.MinValue;
-
-            /// <summary>
-            /// show mass edit over lay for the input duration.
-            /// overrides MassEditOVerlay.Show when it is set to a UTC time in future.
-            /// seconds is
-            /// seconds is
-            /// </summary>
-            /// <param name="seconds"> duration.
-            /// negative => never
-            /// float.MaxValue => always
-            /// </param>
-            public static void SetTimer(float seconds) {
-                DateTime dt;
-                if (seconds == float.MaxValue) {
-                    dt = DateTime.MaxValue;
-                } else {
-                    dt = DateTime.UtcNow + TimeSpan.FromSeconds(seconds);
-                }
-                lock (_lock) {
-                    _timer = dt;
-                }
-            }
-
-            /// <summary>
-            /// show overlay for other subtools influced by mass edit.
-            /// </summary>
-            public static bool IsActive {
-                get {
-                    bool show;
-                    DateTime timer;
-                    lock (_lock) {
-                        show = _show;
-                        timer = _timer;
-                    }
-                    return show || DateTime.UtcNow < timer;
-                }
-            }
-        }
-
-        /// <summary>
         /// refreshes all subtools incflucned by mass edit.
         /// the mass edit overlay active while processing
         /// and remains active for one extra second so that
@@ -230,14 +170,14 @@ namespace TrafficManager.UI.SubTools {
             // that something is happening.
             // this is also to make sure overlay is refresshed
             // even when the user lets go of the mass edit overlay hotkey.
-            MassEditOVerlay.SetTimer(float.MaxValue);
+            MassEditOverlay.SetTimer(float.MaxValue);
 
             ModUI.GetTrafficManagerTool(false)?.InitializeSubTools();
             RefreshCurrentPriorityNodeIds();
 
             // keep active for one more second so that the user
             // has a chance to see the new traffic rules.
-            MassEditOVerlay.SetTimer(1);
+            MassEditOverlay.SetTimer(1);
         }
 
         public override void RenderOverlay(RenderManager.CameraInfo cameraInfo) {
@@ -248,7 +188,7 @@ namespace TrafficManager.UI.SubTools {
             bool ctrlDown = Input.GetKey(KeyCode.LeftControl) || Input.GetKey(KeyCode.RightControl);
             bool shiftDown = Input.GetKey(KeyCode.LeftShift) || Input.GetKey(KeyCode.RightShift);
 
-            MassEditOVerlay.Show = ctrlDown;
+            MassEditOverlay.Show = ctrlDown;
 
             if (ctrlDown) {
                 massEditMode = PrioritySignsMassEditMode.MainYield;
